+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - set `ITSAppUsesNonExemptEncryption` to `false` for mac
 
 ### Fixed
+- fix "message not found in cache" bug #3039
 - fix webxdc: allow `self` and `blob:` in `connect-src` in CSP
 - indentation in update device message
 - fix "no messages" blinking up for a second when loading a chat